# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SHOP Quote API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
components:
  schemas:
...
    RequestUploadUrlBody:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    RequestUploadUrlResponse:
      type: object
      required:
        - uploadURL
        - objectPath
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string